chore(outboundapp): follow the utility package move from #386
#386 moved internal/models/attrs to internal/attrs and internal/models/helpers
to internal/helpers. The two new outbound app model files were written before
that landed and still imported the old paths.
